move manager object into a plugin
