+++ directord/tests/test_main.py
@@ -47,6 +47,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "finger_print": False,
                 "job_port": 5555,
@@ -73,6 +74,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -97,6 +99,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -121,6 +124,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -145,6 +149,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -169,6 +174,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -194,6 +200,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -218,6 +225,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -242,6 +250,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -266,6 +275,7 @@ class TestMain(unittest.TestCase):
                 "check": False,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -289,6 +299,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -311,6 +322,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "transfer_port": 5556,
@@ -331,6 +343,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -358,6 +371,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -385,6 +399,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -412,6 +427,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -439,6 +455,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -466,6 +483,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": "xxxx",
                 "export_nodes": None,
@@ -493,6 +511,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": "xxxx",
@@ -520,6 +539,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "export_jobs": None,
                 "export_nodes": None,
@@ -550,6 +570,7 @@ class TestMain(unittest.TestCase):
                 "config_file": None,
                 "shared_key": None,
                 "curve_encryption": False,
+                "datastore": None,
                 "debug": False,
                 "job_port": 5555,
                 "key_file": None,
